+++ src/schematix/decorators/__init__.py
@@ -1,0 +1,8 @@
+# ~/schematix/src/schematix/decorators/__init__.py
+from .field import field
+from .schema import schema
+
+__all__ = [
+    'field',
+    'schema'
+]